# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShotPilot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type GenerateShotPlanBodyDeviceType =
  (typeof GenerateShotPlanBodyDeviceType)[keyof typeof GenerateShotPlanBodyDeviceType];

export const GenerateShotPlanBodyDeviceType = {
  Mobile: "Mobile",
  DSLR: "DSLR",
  Mirrorless: "Mirrorless",
} as const;

export type GenerateShotPlanBodySkillLevel =
  (typeof GenerateShotPlanBodySkillLevel)[keyof typeof GenerateShotPlanBodySkillLevel];

export const GenerateShotPlanBodySkillLevel = {
  Beginner: "Beginner",
  Intermediate: "Intermediate",
  Pro: "Pro",
} as const;

export interface GenerateShotPlanBody {
  deviceType: GenerateShotPlanBodyDeviceType;
  brandModel: string;
  cameraType: string;
  shootingSituation: string;
  purpose: string;
  skillLevel: GenerateShotPlanBodySkillLevel;
}

export interface ShotPlan {
  cameraAngles: string[];
  shotList: string[];
  cameraSettings: string[];
  movementTips: string[];
  editingSuggestions: string[];
  summary: string;
}

export interface SaveShotPlanBody {
  name: string;
  deviceType: string;
  brandModel: string;
  shootingSituation: string;
  purpose: string;
  skillLevel: string;
  plan: ShotPlan;
}

export interface SavedShotPlan {
  id: number;
  name: string;
  deviceType: string;
  brandModel: string;
  shootingSituation: string;
  purpose: string;
  skillLevel: string;
  plan: ShotPlan;
  createdAt: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}