# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShotPilot API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Generate AI camera shooting plan
 */
export const GenerateShotPlanBody = zod.object({
  deviceType: zod.enum(["Mobile", "DSLR", "Mirrorless"]),
  brandModel: zod.string(),
  cameraType: zod.string(),
  shootingSituation: zod.string(),
  purpose: zod.string(),
  skillLevel: zod.enum(["Beginner", "Intermediate", "Pro"]),
});

export const GenerateShotPlanResponse = zod.object({
  cameraAngles: zod.array(zod.string()),
  shotList: zod.array(zod.string()),
  cameraSettings: zod.array(zod.string()),
  movementTips: zod.array(zod.string()),
  editingSuggestions: zod.array(zod.string()),
  summary: zod.string(),
});

/**
 * @summary List saved shot plans
 */
export const ListSavedPlansResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  deviceType: zod.string(),
  brandModel: zod.string(),
  shootingSituation: zod.string(),
  purpose: zod.string(),
  skillLevel: zod.string(),
  plan: zod.object({
    cameraAngles: zod.array(zod.string()),
    shotList: zod.array(zod.string()),
    cameraSettings: zod.array(zod.string()),
    movementTips: zod.array(zod.string()),
    editingSuggestions: zod.array(zod.string()),
    summary: zod.string(),
  }),
  createdAt: zod.coerce.date(),
});
export const ListSavedPlansResponse = zod.array(ListSavedPlansResponseItem);

/**
 * @summary Save a shot plan
 */
export const SaveShotPlanBody = zod.object({
  name: zod.string(),
  deviceType: zod.string(),
  brandModel: zod.string(),
  shootingSituation: zod.string(),
  purpose: zod.string(),
  skillLevel: zod.string(),
  plan: zod.object({
    cameraAngles: zod.array(zod.string()),
    shotList: zod.array(zod.string()),
    cameraSettings: zod.array(zod.string()),
    movementTips: zod.array(zod.string()),
    editingSuggestions: zod.array(zod.string()),
    summary: zod.string(),
  }),
});

/**
 * @summary Delete a saved plan
 */
export const DeleteSavedPlanParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});