# lib/api-zod/src/generated/types/generateShotPlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShotPilot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateShotPlanBodyDeviceType } from "./generateShotPlanBodyDeviceType";
import type { GenerateShotPlanBodySkillLevel } from "./generateShotPlanBodySkillLevel";

export interface GenerateShotPlanBody {
  deviceType: GenerateShotPlanBodyDeviceType;
  brandModel: string;
  cameraType: string;
  shootingSituation: string;
  purpose: string;
  skillLevel: GenerateShotPlanBodySkillLevel;
}