# lib/api-zod/src/generated/types/saveShotPlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShotPilot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShotPlan } from "./shotPlan";

export interface SaveShotPlanBody {
  name: string;
  deviceType: string;
  brandModel: string;
  shootingSituation: string;
  purpose: string;
  skillLevel: string;
  plan: ShotPlan;
}