# lib/api-zod/src/generated/types/shotPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShotPilot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShotPlan {
  cameraAngles: string[];
  shotList: string[];
  cameraSettings: string[];
  movementTips: string[];
  editingSuggestions: string[];
  summary: string;
}